# cmd/server/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NB_Api",
    "description": "API do CRM Atendimento — Gerenciamento de chamadas e mensagens WhatsApp",
    "version": "1.0.5"
  },
  "servers": [
    { "url": "http://localhost:8081", "description": "Desenvolvimento (dev)" },
    { "url": "http://localhost:8080", "description": "Produção (Docker)" }
  ],
  "security": [
    { "basicAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Credenciais definidas nas variáveis AUTH_USERNAME / AUTH_PASSWORD do .env"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Mensagem de erro" }
        }
      },
      "SessionInfo": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "jid": { "type": "string", "nullable": true },
          "state": { "type": "string", "enum": ["connecting", "qr", "open", "logged_out"] },
          "paired": { "type": "boolean" },
          "apiKey": { "type": "string" }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "messageId": { "type": "string" },
          "timestamp": { "type": "integer" }
        }
      },
...
      "CallIdResponse": {
        "type": "object",
        "properties": {
          "call": {
            "type": "object",
            "properties": {
              "callId": { "type": "string" }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/sessions": {
      "get": {
        "summary": "Listar sessões",
        "tags": ["Sessões"],
        "responses": {
          "200": {
            "description": "Lista de sessões",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SessionInfo" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Criar sessão",
        "tags": ["Sessões"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Nome da instância. O sid será gerado a partir deste nome (slug). Ex: 'Minha Sessao' → sid 'Minha-Sessao'" },
                  "apiKey": { "type": "string", "description": "Chave de API (auto-gerada se vazia)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sessão criada. O campo 'id' é derivado do nome informado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "description": "Session ID (slug do nome)" },
                    "apiKey": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sid}": {
      "delete": {
        "summary": "Excluir sessão",
        "tags": ["Sessões"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Sessão excluída" },
          "404": { "description": "Sessão não encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/sessions/{sid}/logout": {
      "post": {
        "summary": "Desconectar sessão do WhatsApp",
        "tags": ["Sessões"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "Desconectado" },
          "404": { "description": "Sessão não encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/sessions/{sid}/pair": {
      "post": {
        "summary": "Iniciar pareamento (gera QR code)",
        "tags": ["Sessões"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "204": { "description": "QR code gerado — evento session-qr no SSE" },
          "400": { "description": "Sessão já pareada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/sessions/{sid}/apikey": {
      "put": {
        "summary": "Atualizar chave de API da sessão",
        "tags": ["Sessões"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "apiKey": { "type": "string", "description": "Nova chave (auto-gerada se vazia)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chave atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiKey": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sid}/name": {
      "put": {
        "summary": "Alterar nome da sessão",
        "tags": ["Sessões"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Novo nome" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Nome atualizado" }
        }
      }
    },
    "/api/sessions/{sid}/calls": {
      "post": {
        "summary": "Iniciar chamada telefônica",
        "tags": ["Chamadas"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": { "type": "string", "description": "Número de telefone" },
                  "duration_ms": { "type": "integer", "description": "Duração máxima em ms (opcional)" },
                  "record": { "type": "boolean", "description": "Gravar chamada" }
                },
                "required": ["phone"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chamada iniciada",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CallIdResponse" } } }
          },
          "400": { "description": "Phone inválido" },
          "409": { "description": "Operador já em chamada" },
          "429": { "description": "Máximo de chamadas concorrentes" },
          "503": { "description": "Não pareado" }
        }
      }
    },
    "/api/sessions/{sid}/calls/{id}/webrtc": {
      "post": {
        "summary": "Enviar oferta WebRTC",
        "tags": ["Chamadas"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "ID da chamada" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sdp_offer": { "type": "string", "description": "SDP offer do navegador" }
                },
                "required": ["sdp_offer"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SDP answer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sdp_answer": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": { "description": "Chamada não encontrada" }
        }
      }
    },
    "/api/sessions/{sid}/calls/{id}/accept": {
      "post": {
        "summary": "Aceitar chamada recebida",
        "tags": ["Chamadas"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "ID da chamada" }
        ],
        "responses": {
          "200": {
            "description": "Chamada aceita",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CallIdResponse" } } }
          },
          "404": { "description": "Chamada não encontrada" },
          "409": { "description": "Conflito" }
        }
      }
    },
    "/api/sessions/{sid}/calls/{id}/reject": {
      "post": {
        "summary": "Rejeitar chamada recebida",
        "tags": ["Chamadas"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "ID da chamada" }
        ],
        "responses": {
          "200": {
            "description": "Chamada rejeitada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sid}/calls/{id}": {
      "delete": {
        "summary": "Encerrar chamada",
        "tags": ["Chamadas"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "ID da chamada" }
        ],
        "responses": {
          "204": { "description": "Chamada encerrada" }
        }
      }
    },
    "/api/sessions/{sid}/history": {
      "get": {
        "summary": "Histórico de chamadas",
        "tags": ["Chamadas"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Histórico (últimas 50 chamadas)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rows": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/CallRecord" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sid}/messages": {
      "post": {
        "summary": "Enviar mensagem (genérico — dispatcher por type)",
        "tags": ["Mensagens"],
        "description": "O campo `type` define o tipo de mensagem. Os campos adicionais variam conforme o tipo (ver endpoints específicos).",
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["text", "image", "video", "audio", "document", "sticker", "location", "contact", "list"] }
                },
                "required": ["type"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mensagem enviada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "503": { "description": "Não conectado" }
        }
      }
    },
    "/api/sessions/{sid}/messages/text": {
      "post": {
        "summary": "Enviar mensagem de texto",
        "tags": ["Mensagens"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": { "type": "string", "description": "Telefone ou JID do destinatário" },
                  "text": { "type": "string", "description": "Texto da mensagem" }
                },
                "required": ["to", "text"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mensagem enviada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "503": { "description": "Não conectado" }
        }
      }
    },
    "/api/sessions/{sid}/messages/media": {
      "post": {
        "summary": "Enviar mídia (imagem, vídeo, áudio, documento, sticker)",
        "tags": ["Mensagens"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": { "type": "string", "description": "Telefone ou JID" },
                  "type": { "type": "string", "enum": ["image", "video", "audio", "document", "sticker"] },
                  "data": { "type": "string", "format": "byte", "description": "Base64 do arquivo" },
                  "url": { "type": "string", "description": "URL pública (alternativa)" },
                  "caption": { "type": "string" },
                  "filename": { "type": "string", "description": "Nome do arquivo (document)" },
                  "mimetype": { "type": "string" }
                },
                "required": ["to", "type", "data"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Mídia enviada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "503": { "description": "Não conectado" }
        }
      }
    },
    "/api/sessions/{sid}/messages/location": {
      "post": {
        "summary": "Enviar localização",
        "tags": ["Mensagens"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": { "type": "string" },
                  "lat": { "type": "number", "format": "float" },
                  "lng": { "type": "number", "format": "float" },
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "url": { "type": "string" }
                },
                "required": ["to"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Localização enviada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "503": { "description": "Não conectado" }
        }
      }
    },
    "/api/sessions/{sid}/messages/contact": {
      "post": {
        "summary": "Enviar contato (vCard)",
        "tags": ["Mensagens"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": { "type": "string" },
                  "displayName": { "type": "string" },
                  "vcard": { "type": "string", "description": "Dados do vCard" }
                },
                "required": ["to", "vcard"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contato enviado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "503": { "description": "Não conectado" }
        }
      }
    },
...
    "/api/sessions/{sid}/messages/poll": {
      "post": {
        "summary": "Enviar enquete (poll)",
        "tags": ["Mensagens"],
        "description": "Envia uma enquete interativa com múltiplas opções. O destinatário pode votar em uma ou mais opções.",
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "to": { "type": "string", "description": "Telefone ou JID do destinatário" },
                  "name": { "type": "string", "description": "Pergunta da enquete" },
                  "options": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "Opções da enquete (mín. 2)"
                  },
                  "selectableCount": {
                    "type": "integer",
                    "description": "Quantas opções o usuário pode selecionar (padrão: 1)"
                  }
                },
                "required": ["to", "name", "options"],
                "example": {
                  "to": "5531992604940",
                  "name": "Qual departamento você precisa?",
                  "options": ["Vendas", "Suporte", "Financeiro"],
                  "selectableCount": 1
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Enquete enviada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "503": { "description": "Não conectado" }
        }
      }
    },
...
    "/api/sessions/{sid}/contacts": {
      "get": {
        "summary": "Listar todos os contatos",
        "tags": ["Usuários / Contatos"],
        "parameters": [
          { "name": "sid", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Mapa de contatos" }
        }
      }
    },
...
  }
}